Add webpack inspection tools; lazy load DataGrid
