+++ node.api.js
@@ -1,7 +1,8 @@
 const { GenerateSW } = require("workbox-webpack-plugin");
+const { BundleAnalyzerPlugin } = require("webpack-bundle-analyzer");
 
 export default () => ({
-  webpack: (config, { defaultLoaders }) => {
+  webpack: config => {
     // Service Worker
     config.plugins.push(
       new GenerateSW({
@@ -23,6 +24,18 @@ export default () => ({
         ]
       })
     );
+
+    // Bundle analyzer
+    const args = process.argv.slice(3);
+    if (args.includes("--webpack-stats")) {
+      console.log("Building webpack statistics file > stats.json");
+      config.plugins.push(
+        new BundleAnalyzerPlugin({
+          analyzerMode: "disabled",
+          generateStatsFile: true
+        })
+      );
+    }
 
     // Inline SVG loader
     config.module.rules[0].oneOf.unshift({
