+++ package.json
@@ -9,7 +9,9 @@
     "serve": "serve dist -p 3000",
     "lint": "eslint ./src/**/*.js ./src/**/*.jsx",
     "storybook": "export NODE_PATH=src/:.storybook/ && start-storybook -p 9009 -s public",
-    "build-storybook": "export NODE_PATH=src/:.storybook/ && build-storybook -c .storybook -s public"
+    "build-storybook": "export NODE_PATH=src/:.storybook/ && build-storybook -c .storybook -s public",
+    "webpack-build": "react-static build --webpack-stats",
+    "webpack-analyze": "webpack-bundle-analyzer dist/stats.json"
   },
   "dependencies": {
     "@babel/core": "^7.5.5",
@@ -78,6 +80,7 @@
     "eslint-plugin-prettier": "^3.1.0",
     "eslint-plugin-react": "7.x",
     "prettier": "^1.18.2",
-    "serve": "10.1.2"
+    "serve": "10.1.2",
+    "webpack-bundle-analyzer": "^3.4.1"
   }
 }
