+++ src/components/DataGrid/AddRowModal/style.scss
@@ -1,0 +1,13 @@
+.add-row-modal {
+    .modal-header {
+        .h4 {
+            margin-bottom: 0;
+            line-height: 1.15;
+        }
+    }
+
+    .modal-body {
+        padding-left: 2rem;
+        padding-right: 2rem;
+    }
+}
