+++ src/components/Icon/fontawesome.js
@@ -1,14 +1,18 @@
-export {
-  faChevronRight,
-  faChevronLeft,
-  faPlus,
-  faSignOutAlt,
-  faMoon,
-  faSun,
-  faShieldAlt,
-  faFilter,
-  faTimesCircle,
-  faQuestionCircle,
-  faMinus
-} from "@fortawesome/free-solid-svg-icons";
-export { faDiscord, faGithub } from "@fortawesome/free-brands-svg-icons";
+/* eslint-disable prettier/prettier */
+
+// Solid icons
+export { faChevronRight } from "@fortawesome/free-solid-svg-icons/faChevronRight";
+export { faChevronLeft } from "@fortawesome/free-solid-svg-icons/faChevronLeft";
+export { faPlus } from "@fortawesome/free-solid-svg-icons/faPlus";
+export { faSignOutAlt } from "@fortawesome/free-solid-svg-icons/faSignOutAlt";
+export { faMoon } from "@fortawesome/free-solid-svg-icons/faMoon";
+export { faSun } from "@fortawesome/free-solid-svg-icons/faSun";
+export { faShieldAlt } from "@fortawesome/free-solid-svg-icons/faShieldAlt";
+export { faFilter } from "@fortawesome/free-solid-svg-icons/faFilter";
+export { faTimesCircle } from "@fortawesome/free-solid-svg-icons/faTimesCircle";
+export { faQuestionCircle } from "@fortawesome/free-solid-svg-icons/faQuestionCircle";
+export { faMinus } from "@fortawesome/free-solid-svg-icons/faMinus";
+
+// Brand icons
+export { faDiscord } from "@fortawesome/free-brands-svg-icons/faDiscord";
+export { faGithub } from "@fortawesome/free-brands-svg-icons/faGithub";
